[UNI-21] refac : 타입 분리, 코드 정리
