+++ uniro_frontend/src/constant/enum/markerEnum.ts
@@ -8,8 +8,3 @@ export const enum Markers {
 	WAYPOINT = "waypoint",
 	NUMBERED_WAYPOINT = "numberedWayPoint",
 }
-
-export const enum RoutePoint {
-	ORIGIN = "origin",
-	DESTINATION = "destination",
-}
