+++ uniro_frontend/src/constant/enum/routeEnum.ts
@@ -1,0 +1,4 @@
+export const enum RoutePoint {
+	ORIGIN = "origin",
+	DESTINATION = "destination",
+}
