+++ uniro_frontend/src/pages/map.tsx
@@ -15,8 +15,8 @@ import useSearchBuilding from "../hooks/useSearchBuilding";
 import Button from "../components/customButton";
 import { AdvancedMarker, MarkerTypes } from "../data/types/marker";
 import { RoutePointType } from "../data/types/route";
-import { RoutePoint } from "../constant/enums";
-import { Markers } from "../constant/enums";
+import { RoutePoint } from "../constant/enum/routeEnum";
+import { Markers } from "../constant/enum/markerEnum";
 import createAdvancedMarker from "../utils/markers/createAdvanedMarker";
 
 export type SelectedMarkerTypes = {
@@ -52,7 +52,7 @@ export default function MapPage() {
 
 					if (type === Markers.BUILDING) return undefined;
 
-					element.content = createMarkerElement({ type, });
+					element.content = createMarkerElement({ type });
 				}
 				return undefined;
 			});
